+++ erde/compile/constants.lua
@@ -26,8 +26,7 @@ CC.KEYWORDS = {
   'continue',
 }
 
--- Words that are keywords in Lua but NOT in Erde. These are allowed to be used
--- as variable names in Erde, but must be transformed when compiling.
+-- Words that are keywords in Lua but NOT in Erde.
 CC.LUA_KEYWORDS = {
   ['not'] = true,
   ['and'] = true,